Enable publishing user-guide by travis

+++ src/build/publish-doc.sh
@@ -1,9 +1,10 @@
 #!/bin/bash -x
 set -e # exit with nonzero exit code if anything fails
 
-
-# DISABLE document publishing for now
-exit 0
+# if [ "$TRAVIS_BRANCH" != 'master' ] || [ "$TRAVIS_PULL_REQUEST" == 'true' ]; then
+if [ "$TRAVIS_PULL_REQUEST" == 'true' ]; then
+    exit 0
+fi
 
 cd $HOME
 rm -fr gh-pages
@@ -13,10 +14,9 @@ git config --global user.name "travis-ci"
 git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/ttddyy/datasource-proxy gh-pages > /dev/null
 
 cd gh-pages
-rm -fr docs/datasource-proxy
-rm -fr docs/datasource-test-proxy
-cp -Rf $TRAVIS_BUILD_DIR/datasource-proxy/target/generated-docs docs/datasource-proxy
-cp -Rf $TRAVIS_BUILD_DIR/datasource-test-proxy/target/generated-docs docs/datasource-test-proxy
+./mvnw asciidoctor:process-asciidoc@output-html
+rm -fr docs/current/user-guide
+cp -Rf $TRAVIS_BUILD_DIR/datasource-proxy/target/generated-docs docs/current/user-guide
 git add -f .
 git commit -m "Latest documentation on successful travis build $TRAVIS_BUILD_NUMBER auto-pushed to gh-pages"
 git push -fq origin gh-pages > /dev/null
